filtering now works
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/infovis/trunk@49 7a0b5368-9e9f-4da2-a48c-2b820cea8ffb

+++ src/FlagStorage.h
@@ -40,7 +40,7 @@ namespace infovis {
 			, SOFTSELECTED = 1 << 3
 		};
 
-		typedef uint8_t FlagItemType;
+		typedef uint32_t FlagItemType;
 
 		typedef std::vector<FlagItemType> FlagVectorType;
